+++ to-do-api-rest/src/main/java/com/todo_ec/model/repository/TodoTareaRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TodoTareaRepository extends CrudRepository <TodoTarea, Long> {
+public interface TodoTareaRepository extends CrudRepository <TodoTarea, Integer> {
 }
